requirements.txt
albumentations==1.3.1
hydra-core==1.3.2
numpy==1.25.1
omegaconf==2.3.0
opencv-python==4.8.0.74
panopticapi @ git+https://github.com/cocodataset/panopticapi.git@7bb4655548f98f3fedc07bf37e9040a992b054b0
phenobench==0.1.0
Pillow==10.0.0
pycocotools==2.0.6
lightning==2.0.5
torch==2.0.1
torch-ema==0.3
torchmetrics==0.10.3
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
wandb==0.15.5